Move enclosingFunction to expressions module
