+++ frontend/src/metabase-lib/v1/expressions/index.ts
@@ -4,3 +4,4 @@ export * from "./identifier";
 export * from "./formatter";
 export * from "./types";
 export * from "./diagnostics";
+export * from "./position";
