+++ frontend/src/metabase-lib/v1/expressions/position.unit.spec.ts
@@ -1,4 +1,4 @@
-import { enclosingFunction } from "./utils";
+import { enclosingFunction } from "./position";
 
 describe("enclosingFunction", () => {
   function setup(doc: string) {
